hw1/requirements.txt
matplotlib==3.10.1
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
seaborn==0.13.2
torch==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
